Add option to ignore checksum check failures
In some scenarios, ignoring checksum failures can be desireable to partially or totally (if the damaged segment landed on unallocated space in the filesystem) recover data from an image.

This lets the user use the `-i`/`--ignore-checksum` option to only display a warning when a checksum fails, but not abort the restore process.