Avoid small buffer overflows computing xxhash
The reason for the hexdigest.resize call was to remove spurious
appended characters from a reading buffer overflow.
Terminate the digest_chars string to avoid such overflow.
Note that is not a security concern, zeroes are likely in
memory and we would just read some more bytes to then discard
them.

Signed-off-by: Frediano Ziglio <frediano.ziglio@cloud.com>